Move BpmnQuerySelectors to a dedicated file.
bundles tests need it and if it stays in bpmn-elements-registry.ts, mxgraph is
loaded and would require JSDOM. We only want to test what is happening in the
html page, so we don't want to manage everything on the NodeJS Jest runner side.
Putting the class in a dedicated file remove mxgraph loading and everything is
working smoothly.

# Conflicts:
#	src/component/registry/bpmn-elements-registry.ts
