+++ src/component/registry/bpmn-elements-registry.ts
@@ -20,6 +20,7 @@ import { FlowKind } from '../../model/bpmn/internal/edge/FlowKind';
 import { ShapeBpmnElementKind } from '../../model/bpmn/internal/shape';
 import { CssRegistry } from './css-registry';
 import MxGraphCellUpdater from '../mxgraph/MxGraphCellUpdater';
+import { BpmnQuerySelectors } from './query-selectors';
 
 export function newBpmnElementsRegistry(graph: BpmnMxGraph): BpmnElementsRegistry {
   return new BpmnElementsRegistry(
@@ -215,60 +216,6 @@ class BpmnModelRegistry {
     }
 
     return { id: bpmnElementId, name: mxCell.value, isShape: mxCell.isVertex(), kind: extractBpmnKindFromStyle(mxCell) };
-  }
-}
-
-/**
- * Once mxGraph is initialized at BpmnVisualization construction, prior loading a BPMN diagram, the DOM looks like:
- * ```html
- * <div id="bpmn-container" style="touch-action: none;">
- *   <svg style="left: 0px; top: 0px; width: 100%; height: 100%; display: block; min-width: 1px; min-height: 1px;">
- *     <g>
- *       <g></g>
- *       <g></g>
- *       <g></g>
- *       <g></g>
- *     </g>
- *   </svg>
- * </div>
- * ```
- *
- * After loading, the DOM looks like:
- * ```html
- * <div id="bpmn-container" style="touch-action: none;">
- *   <svg style="left: 0px; top: 0px; width: 100%; height: 100%; display: block; min-width: 900px; min-height: 181px;">
- *     <g>
- *       <g></g>
- *       <g>
- *         <g style="" class="bpmn-pool" data-bpmn-id="Participant_1">....</g>
- *       </g>
- *       <g></g>
- *       <g></g>
- *     </g>
- *   </svg>
- * </div>
- * ```
- * In the 2nd 'g' node, children 'g' nodes with the 'data-cell-id' attribute (extra attribute generated by the lib) are only available when the rendering is done
- */
-export class BpmnQuerySelectors {
-  constructor(private containerId: string) {}
-
-  // TODO do we make explicit that this selector targets a SVG group?
-  existingElement(): string {
-    return `#${this.containerId} > svg > g > g > g[data-bpmn-id]`;
-  }
-
-  element(bpmnElementId: string): string {
-    // TODO use more precise selector (use child combinator)
-    return `#${this.containerId} svg g g[data-bpmn-id="${bpmnElementId}"]`;
-  }
-
-  labelOfElement(bpmnElementId?: string): string {
-    return `#${this.containerId} > svg > g > g > g[data-bpmn-id="${bpmnElementId}"].bpmn-label > g > foreignObject`;
-  }
-
-  elementsOfKind(bpmnKindCssClassname: string): string {
-    return `#${this.containerId} > svg > g > g > g.${bpmnKindCssClassname}:not(.bpmn-label)`;
   }
 }
